chore: moved tests close to the files they are testing following good practices. Fixed basic test for compare page and created mocks for testing
