--- app/main/character-list.tsx.old
@@ -1,29 +1,0 @@
-import { ICharacterList } from '@/utils/types';
-import Image from 'next/image'
-import Link from 'next/link'
-
-// SEB: Will this component be used elsewhere? It feels like a slightly unnecessary component at the moment as it couple be implemented as a couple of lines in the neighbouring page.tsx component.
-
-export default function CharacterList({ characters }: ICharacterList ) {
-  return (<>
-    {characters && (
-      <ul className="flex m-auto w-4/5 flex-wrap justify-center">
-        {characters.map((character) => (
-          <li key={character.name} >
-            <Link href={`/[${character.name}].tsx`}>
-                  <Image src={`/character/${character.name}.png`}
-                    height={144}
-                    width={144}
-                    alt="Character-Image"
-                    className="w-auto mb-4 rounded shadow-md"
-                  />
-               </Link>
-            </li>
-          ))}
-        </ul>
-      )}
-    </>
-  );
-}
-
-
